Documentación/botcon/requirements.txt
PyPDF2==2.3.1
reportlab==3.6.10
pandas==1.4.2
psycopg2==2.9.3
selenium
webdriver_manager
pywin32
pygsheets
numpy
python-docx
plyer 